+++ client/package.json
@@ -12,15 +12,18 @@
   },
   "author": "Simon Friis Vindum",
   "license": "MIT",
+  "dependencies": {
+    "dffptch": "^1.0.4",
+    "underscore": "^1.8.3",
+    "sync-promise": "^1.0.1",
+    "minivents": "^1.2.0"
+  },
   "devDependencies": {
     "babel-core": "^6.7.6",
     "babel-loader": "^6.2.4",
     "babel-plugin-transform-remove-console": "^6.5.0",
     "babel-preset-es2015": "^6.4.0",
-    "dffptch": "^1.0.4",
-    "minivents": "^1.2.0",
     "sinon": "^2.0.0-pre",
-    "sync-promise": "^1.0.1",
     "testem": "^1.6.0",
     "uglifyjs": "^2.4.10",
     "webpack": "^1.13.0"
