+++ client/webpack.config.js
@@ -4,6 +4,11 @@ const isProduction = process.env.NODE_ENV === 'production';
 const isAmd = process.env.LIBRARY_TARGET === 'amd';
 const outputFileName = `dist/synceddb${isAmd ? '-amd' : '-global'}${isProduction ? '.min' : ''}.js`;
 
+const plugins = [];
+if (isProduction) {
+  plugins.push('transform-remove-console');
+}
+
 module.exports = {
   entry: ['./synceddb.js'],
   output: {
@@ -20,7 +25,7 @@ module.exports = {
         exclude: /(node_modules|bower_components)/,
         query: {
           presets: ['es2015'],
-          plugins: [ 'transform-remove-console' ],
+          plugins,
         },
       },
     ]
